+++ dal/helpers/parsers.py
@@ -16,6 +16,12 @@ from movai_core_shared.logger import Log
 from dal.models.scopestree import scopes
 from dal.models.var import Var
 from dal.movaidb import MovaiDB
+from dal.exceptions import (
+    UndefinedFlowParameterError,
+    UndefinedConfigParameterError,
+    UndefinedVarParameterError,
+    UndefinedParamParameterError,
+)
 
 if TYPE_CHECKING:
     from dal.models.container import Container
@@ -179,7 +185,9 @@ class ParamParser:
             obj = cast("Configuration", scopes.from_path(_config_name, scope="Configuration"))
 
         except KeyError as exc:
-            raise ValueError(f"Configuration {_config_name} does not exist") from exc
+            raise UndefinedConfigParameterError(
+                f"Configuration {_config_name} does not exist"
+            ) from exc
 
         output = obj.get_param(_config_param)
 
@@ -208,9 +216,17 @@ class ParamParser:
 
         cls_name = type(instance).__name__
         if cls_name == "Flow":  # Flows don't have a node name
+            if not instance.has_param(param_name):
+                raise UndefinedParamParameterError(
+                    f'Parameter "{param_name}" is not defined in flow "{instance.ref}"'
+                )
             instance = cast("Flow", instance)
             output = instance.get_param(param_name, self.context) or default
         elif cls_name in ["NodeInst", "Container"]:
+            if not instance.has_param(param_name, node_name, self.context):
+                raise UndefinedParamParameterError(
+                    f'Parameter "{param_name}" is not defined in "{node_name}" of flow "{instance.flow.ref}"'
+                )
             output = instance.get_param(param_name, node_name, self.context) or default
         else:
             raise ValueError(f'Instance type "{cls_name}" not supported')
@@ -238,7 +254,7 @@ class ParamParser:
         output = Var(context, robot_name).get(param_name)
 
         if not output:
-            raise ValueError(f'"{param_name}" does not exist in Var "{context}"')
+            raise UndefinedVarParameterError(f'"{param_name}" does not exist in Var "{context}"')
 
         return output
 
@@ -266,38 +282,14 @@ class ParamParser:
         node_name_arr = node_name.split("__")
         # Check if this is the main flow or a subflow
         is_subflow = len(node_name_arr) > 1
-        value = instance.flow.get_param(param_name, context=self.context, is_subflow=is_subflow)
-        if value is None:
-            value = default
 
-        if len(node_name_arr) > 1:
-            # instance is not in the main flow
+        flow = instance.flow
+        if not flow.has_param(param_name):
+            raise UndefinedFlowParameterError(
+                f'Flow parameter "{param_name}" is not defined in flow "{flow.ref}"'
+            )
 
-            # not using istance bc import
-            if type(instance).__name__ in ["NodeInst", "Container"]:
-                ctr_arr = node_name_arr[:-1]
-
-                if ctr_arr:
-                    # get the name of the container
-                    _name = "__".join(ctr_arr)
-
-                    # get the container instance
-                    ctr_instance = self.flow.get_container(_name, self.context)
-                    assert ctr_instance is not None, f"Container {_name} not found"
-
-                    # get the instance parameter value
-                    # if there is no instance param, set to default
-                    ctr_value = ctr_instance.get_param(
-                        param_name, _name, self.context, default_value=value
-                    )
-
-                    value = value if ctr_value is None else ctr_value
-
-            else:
-                msg = f'Instance type "{type(instance).__name__}" not supported'
-                raise ValueError(msg)
-
-        return value
+        return flow.get_param(param_name, context=self.context, is_subflow=is_subflow)
 
 
 def get_string_from_template(template: str, task_entry: object) -> str: